+++ NEWS.md
@@ -30,6 +30,7 @@
 - Replace `as.character.quosure` usage with `as_label` due to deprecation
 - Minor changes to the exact calculation of the anchor weight matrix
 - Default rasterization limit in `DimPlot()` and `FeaturePlot()` changed from 50,000 to 100,000
+- `SCTransform()` now returns a formalized `Assay` subclass `SCTAssay()`
 
 ## Removed
 - `CreateGeneActivityMatrix` replaced by `Signac::GeneActivity()`